# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/practice/sessions": {
      "post": {
        "operationId": "PracticeController_createSession",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Practice session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSession"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new practice session",
        "tags": [
          "Practice"
        ]
      }
    },
    "/api/v1/practice/sessions/{id}": {
      "get": {
        "operationId": "PracticeController_getSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSession"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get practice session by ID",
        "tags": [
          "Practice"
        ]
      }
    },
    "/api/v1/practice/sessions/{id}/answers": {
      "post": {
        "operationId": "PracticeController_submitAnswer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Answer submitted"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Submit an answer for a question",
        "tags": [
          "Practice"
        ]
      }
    },
    "/api/v1/practice/sessions/{id}/complete": {
      "post": {
        "operationId": "PracticeController_completeSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session completed"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Complete a practice session",
        "tags": [
          "Practice"
        ]
      }
    },
    "/api/v1/practice/sessions/{id}/results": {
      "get": {
        "operationId": "PracticeController_getResults",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Results retrieved"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get practice session results",
        "tags": [
          "Practice"
        ]
      }
    },
    "/api/v1/questions": {
      "get": {
        "operationId": "QuestionsController_getQuestions",
        "parameters": [
          {
            "name": "level",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "section",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questions retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Question"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get questions by level and section",
        "tags": [
          "Questions"
        ]
      }
    },
    "/api/v1/questions/random": {
      "get": {
        "operationId": "QuestionsController_getRandomQuestions",
        "parameters": [
          {
            "name": "level",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "section",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Random questions retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Question"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get random questions by level and section",
        "tags": [
          "Questions"
        ]
      }
    },
    "/api/v1/questions/{id}": {
      "get": {
        "operationId": "QuestionsController_getQuestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get question by ID",
        "tags": [
          "Questions"
        ]
      }
    },
    "/api/v1/questions/stats": {
      "get": {
        "operationId": "QuestionsController_getStats",
        "parameters": [
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stats retrieved"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get question statistics",
        "tags": [
          "Questions"
        ]
      }
    },
    "/api/v1/exams": {
      "get": {
        "operationId": "ExamsController_getExams",
        "parameters": [
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exams retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exam"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List published exams",
        "tags": [
          "Exams"
        ]
      }
    },
    "/api/v1/exams/{id}": {
      "get": {
        "operationId": "ExamsController_getExamById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exam retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get exam by ID",
        "tags": [
          "Exams"
        ]
      }
    },
    "/api/v1/exams/{id}/submit": {
      "post": {
        "operationId": "ExamsController_submitExam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitExamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exam submitted successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Submit exam answers",
        "tags": [
          "Exams"
        ]
      }
    }
  },
  "info": {
    "title": "TacoLearn API",
    "description": "Backend API template for the TacoLearn application.\n\n## Response Format\nAll responses follow a standardized format:\n- **status**: HTTP status code\n- **message**: Human-readable message\n- **data**: Response data (object or array)\n- **pagination**: (Optional) Pagination metadata for list endpoints",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "JlptLevel": {
        "type": "string",
        "enum": [
          "N1",
          "N2",
          "N3",
          "N4",
          "N5"
        ]
      },
      "QuestionSection": {
        "type": "string",
        "enum": [
          "VOCABULARY",
          "GRAMMAR",
          "KANJI",
          "READING",
          "LISTENING"
        ]
      },
      "QuestionType": {
        "type": "string",
        "enum": [
          "MULTIPLE_CHOICE",
          "FILL_IN",
          "MATCHING"
        ]
      },
      "Difficulty": {
        "type": "string",
        "enum": [
          "EASY",
          "MEDIUM",
          "HARD"
        ]
      },
      "QuestionSource": {
        "type": "string",
        "enum": [
          "OFFICIAL_PAST_PAPER",
          "TEXTBOOK",
          "AI_GENERATED",
          "MANUAL"
        ]
      },
...
      "SessionType": {
        "type": "string",
        "enum": [
          "TARGETED",
          "MIXED",
          "WEAK_AREAS",
          "PREVIOUS_MISTAKES",
          "MOCK_TEST"
        ]
      },
      "SessionStatus": {
        "type": "string",
        "enum": [
          "IN_PROGRESS",
          "COMPLETED",
          "ABANDONED"
        ]
      },
      "PracticeSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          },
          "student": {
            "$ref": "#/components/schemas/StudentProfile"
          },
          "sessionType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionType"
              }
            ]
          },
          "examId": {
            "type": "string",
            "nullable": true
          },
          "exam": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Exam"
              }
            ]
          },
          "jlptLevel": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/JlptLevel"
              }
            ]
          },
          "skill": {
            "type": "string",
            "nullable": true
          },
          "totalQuestions": {
            "type": "integer",
            "format": "int32"
          },
          "questionsCompleted": {
            "type": "integer",
            "format": "int32"
          },
          "correctAnswers": {
            "type": "integer",
            "format": "int32"
          },
          "totalTimeSeconds": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionStatus"
              }
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAttempt"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "studentId",
          "sessionType",
          "examId",
          "jlptLevel",
          "skill",
          "totalQuestions",
          "questionsCompleted",
          "correctAnswers",
          "totalTimeSeconds",
          "status",
          "startedAt",
          "completedAt",
          "createdAt"
        ]
      },
      "Exam": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "jlptLevel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/JlptLevel"
              }
            ]
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExamType"
              }
            ]
          },
          "year": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "totalQuestions": {
            "type": "integer",
            "format": "int32"
          },
          "isPublished": {
            "type": "boolean"
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PracticeSession"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "jlptLevel",
          "type",
          "year",
          "month",
          "durationMinutes",
          "totalQuestions",
          "isPublished",
          "createdById",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "jlptLevel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/JlptLevel"
              }
            ]
          },
          "section": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionSection"
              }
            ]
          },
          "skill": {
            "type": "string"
          },
          "questionType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionType"
              }
            ]
          },
          "difficulty": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Difficulty"
              }
            ]
          },
          "content": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "choices": {
            "type": "object"
          },
          "correctAnswerCode": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "difficultyIndex": {
            "type": "string",
            "nullable": true,
            "format": "Decimal.js"
          },
          "attemptCount": {
            "type": "integer",
            "format": "int32"
          },
          "correctCount": {
            "type": "integer",
            "format": "int32"
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionSource"
              }
            ]
          },
          "sourceReference": {
            "type": "string",
            "nullable": true
          },
          "examId": {
            "type": "string",
            "nullable": true
          },
          "exam": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Exam"
              }
            ]
          },
          "orderInExam": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAttempt"
            }
          },
          "vocabulary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vocabulary"
            }
          },
          "grammarPoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrammarPoint"
            }
          },
          "lessons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lesson"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "jlptLevel",
          "section",
          "skill",
          "questionType",
          "difficulty",
          "content",
          "imageUrl",
          "choices",
          "correctAnswerCode",
          "explanation",
          "tags",
          "difficultyIndex",
          "attemptCount",
          "correctCount",
          "source",
          "sourceReference",
          "examId",
          "orderInExam",
          "createdById",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "QuestionAttempt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          },
          "student": {
            "$ref": "#/components/schemas/StudentProfile"
          },
          "questionId": {
            "type": "string"
          },
          "question": {
            "$ref": "#/components/schemas/Question"
          },
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "session": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PracticeSession"
              }
            ]
          },
          "studentAnswerCode": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "timeSpentSeconds": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "confidenceLevel": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ConfidenceLevel"
              }
            ]
          },
          "hintUsed": {
            "type": "boolean"
          },
          "markedForReview": {
            "type": "boolean"
          },
          "mistakeType": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/MistakeType"
              }
            ]
          },
          "attemptedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "studentId",
          "questionId",
          "sessionId",
          "studentAnswerCode",
          "isCorrect",
          "timeSpentSeconds",
          "confidenceLevel",
          "hintUsed",
          "markedForReview",
          "mistakeType",
          "attemptedAt"
        ]
      },
...
      "StudentProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "jlptGoalLevel": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/JlptLevel"
              }
            ]
          },
          "jlptGoalDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "currentReadinessScore": {
            "type": "string",
            "format": "Decimal.js"
          },
          "totalPracticeTimeHours": {
            "type": "string",
            "format": "Decimal.js"
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAttempt"
            }
          },
          "skillMastery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillMastery"
            }
          },
          "srsItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SrsItem"
            }
          },
          "dailyPlans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyLearningPlan"
            }
          },
          "practiceSessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PracticeSession"
            }
          },
          "enrollments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassEnrollment"
            }
          },
          "attendances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attendance"
            }
          },
          "lessonProgress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonProgress"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "jlptGoalLevel",
          "jlptGoalDate",
          "currentReadinessScore",
          "totalPracticeTimeHours",
          "bio",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "emailVerifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "profile": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserProfile"
              }
            ]
          },
          "studentProfile": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/StudentProfile"
              }
            ]
          },
          "teacherProfile": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TeacherProfile"
              }
            ]
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "schoolsOwned": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/School"
            }
          },
          "examsCreated": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Exam"
            }
          },
          "questionsCreated": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          },
          "lessonsCreated": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lesson"
            }
          },
          "announcementsCreated": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Announcement"
            }
          },
          "lastLoginAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "isActive",
          "emailVerifiedAt",
          "lastLoginAt",
          "deletedAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "SubmitExamDto": {
        "type": "object",
        "properties": {
          "answers": {
            "description": "Answers for each question",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerDto"
            }
          },
          "timeSpent": {
            "type": "number",
            "description": "Time spent in seconds"
          }
        },
        "required": [
          "answers",
          "timeSpent"
        ]
      }
    }
  }
}